categories

+++ admin/manage-user.php
@@ -13,15 +13,53 @@ $users = mysqli_query($con, $query);
     }
 </style>
 <section class="dashboard">
-    <?php if (isset($_SESSION['add-user-sucess'])) : ?>
+    <?php if (isset($_SESSION['add-user-success'])) : //show if adding user successful 
+    ?>
         <div class="alert__message success container">
             <p>
                 <?= $_SESSION['add-user-success'];
-                unset($_SESSION['add-user-success'])
+                unset($_SESSION['add-user-success']);
+                ?>
+            </p>
+        </div>
+    <?php elseif (isset($_SESSION['edit-user-success'])) : //show if edit user successful 
+    ?>
+        <div class="alert__message success container">
+            <p>
+                <?= $_SESSION['edit-user-success'];
+                unset($_SESSION['edit-user-success']);
+                ?>
+            </p>
+        </div>
+    <?php elseif (isset($_SESSION['edit-user'])) : //show if edit user NOT successful 
+    ?>
+        <div class="alert__message error container">
+            <p>
+                <?= $_SESSION['edit-user'];
+                unset($_SESSION['edit-user']);
+                ?>
+            </p>
+        </div>
+    <?php elseif (isset($_SESSION['delete-user'])) : //show if delete user NOT successful 
+    ?>
+        <div class="alert__message error container">
+            <p>
+                <?= $_SESSION['delete-user'];
+                unset($_SESSION['delete-user']);
+                ?>
+            </p>
+        </div>
+    <?php elseif (isset($_SESSION['delete-user-success'])) : //show if delete user successful 
+    ?>
+        <div class="alert__message success container">
+            <p>
+                <?= $_SESSION['delete-user-success'];
+                unset($_SESSION['delete-user-success']);
                 ?>
             </p>
         </div>
     <?php endif ?>
+
     <div class="container dashboard__container">
         <button id="show__sidebar-btn" class="sidebar__toggle"><i class="uil uil-angle-right-b"></i></button>
         <button id="hide__sidebar-btn" class="sidebar__toggle"><i class="uil uil-angle-left-b"></i></button>
@@ -68,28 +106,32 @@ $users = mysqli_query($con, $query);
         </aside>
         <main>
             <h2>Manage Users</h2>
-            <table>
-                <thead>
-                    <tr>
-                        <th>Name</th>
-                        <th>Edit</th>
-                        <th>Delete</th>
-                        <th>Admin</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <?php while ($user = mysqli_fetch_assoc($users)) : ?>
+            <?php if (mysqli_num_rows($users) > 0) : ?>
+                <table>
+                    <thead>
                         <tr>
-                            <td><?= "{$user['firstname']} {$user['lastname']}" ?> </td>
-                            <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?= $user['id'] ?> " class="btn sm edit">Edit</a></td>
-                            <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?= $user['id'] ?>" class="btn sm danger">Delete</a></td>
-                            <td><?= $user['is_admin'] ? 'Yes' : 'No' ?></td>
+                            <th>Name</th>
+                            <th>Edit</th>
+                            <th>Delete</th>
+                            <th>Admin</th>
                         </tr>
-                    <?php endwhile ?>
+                    </thead>
+                    <tbody>
+                        <?php while ($user = mysqli_fetch_assoc($users)) : ?>
+                            <tr>
+                                <td><?= "{$user['firstname']} {$user['lastname']}" ?> </td>
+                                <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?= $user['id'] ?> " class="btn sm edit">Edit</a></td>
+                                <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?= $user['id'] ?>" class="btn sm danger">Delete</a></td>
+                                <td><?= $user['is_admin'] ? 'Yes' : 'No' ?></td>
+                            </tr>
+                        <?php endwhile ?>
 
 
-                </tbody>
-            </table>
+                    </tbody>
+                </table>
+            <?php else : ?>
+                <div class="alert__message error"><?= "No users found." ?></div>
+            <?php endif ?>
         </main>
     <?php endif ?>
